fix(rootfs): set fs.may_detach_mounts to 1 and go mod tidy
